[squash] defer responsibility of unlocking buckets (i.e. removing from bucket queue in zookeeper) to the object queue consumer
The idea behind is to avoid queueing object delete entries multiple
times, in case the time required to delete objects from a bucket is
greater than the delay between two bucket processing tasks (default
5mn). It does not prevent the case to happen but should make it very
unlikely in practice since the bucket will be unlocked while the last
objects are being deleted. Note that a strong guarantee is not
required here, so we should be fine.

Also fix an issue with the last modified check in consumer (check
error object)